# bpo-44889: Specialize LOAD_METHOD with PEP 659 adaptive interpreter