--- a/Include/opcode.h
+++ b/Include/opcode.h
@@ -149,10 +149,12 @@ extern "C" {
 #define LOAD_GLOBAL_ADAPTIVE     41
 #define LOAD_GLOBAL_MODULE       42
 #define LOAD_GLOBAL_BUILTIN      43
-#define STORE_ATTR_ADAPTIVE      44
-#define STORE_ATTR_SPLIT_KEYS    45
-#define STORE_ATTR_SLOT          46
-#define STORE_ATTR_WITH_HINT     47
+#define LOAD_METHOD_ADAPTIVE     44
+#define LOAD_METHOD_WITH_HINT    45
+#define STORE_ATTR_ADAPTIVE      46
+#define STORE_ATTR_SPLIT_KEYS    47
+#define STORE_ATTR_SLOT          48
+#define STORE_ATTR_WITH_HINT     58
 #ifdef NEED_OPCODE_JUMP_TABLES
 static uint32_t _PyOpcode_RelativeJump[8] = {
     0U,