--- a/Lib/opcode.py
+++ b/Lib/opcode.py
@@ -234,7 +234,6 @@ _specialized_instructions = [
     "LOAD_GLOBAL_MODULE",
     "LOAD_GLOBAL_BUILTIN",
     "LOAD_METHOD_ADAPTIVE",
-    "LOAD_METHOD_SPLIT_KEYS",
     "LOAD_METHOD_WITH_HINT",
 ]