--- a/Python/specialize.c
+++ b/Python/specialize.c
@@ -121,6 +121,7 @@ _Py_GetSpecializationStats(void) {
     int err = 0;
     err += add_stat_dict(stats, LOAD_ATTR, "load_attr");
     err += add_stat_dict(stats, LOAD_GLOBAL, "load_global");
+    err += add_stat_dict(stats, LOAD_GLOBAL, "load_method");
     err += add_stat_dict(stats, BINARY_SUBSCR, "binary_subscr");
     if (err < 0) {
         Py_DECREF(stats);