Up to 24% faster in `logging_silent`, 2% faster on average in pyperformance.

<!-- issue-number: [bpo-44889](https://bugs.python.org/issue44889) -->
https://bugs.python.org/issue44889
<!-- /issue-number -->