@Fidget-Spinner 2021-08-11
> I'd be interested in seeing the stats for a run on pyperformance, or some other reasonably sized code.

I don't have pyperformance with me at the moment, but here's a subset of the test suite:

```
python.exe -m test test_typing test_re test_dis test_zlib

    load_method.specialization_success : 1271
    load_method.specialization_failure : 1400
    load_method.hit : 918753
    load_method.deferred : 93471
    load_method.miss : 1425
    load_method.deopt : 183
    load_method.unquickened : 8606
    ...
    load_method.specialization_failure_kinds[15] : 0 (is attribute, common in other groups of tests)
    load_method.specialization_failure_kinds[16] : 0 (dict subclass, note: only 2 tests in the entire test suite managed to trip this)
    load_method.specialization_failure_kinds[17] : 415 (module method)
    load_method.specialization_failure_kinds[18] : 416 (class method)
    load_method.specialization_failure_kinds[19] : 132 (not method)
```
FWIW, that's better than what load/store_attr gets. I think this shows that majority of the cases are ones with a real method, and also explains why the original `LOAD_METHOD` + `CALL_METHOD` moved pyperformance.