@markshannon 2021-08-11
Those numbers seem to roughly match https://github.com/faster-cpython/ideas/issues/81#issuecomment-896059763
I wonder if my "mysteriously missing" ~7% correspond to your `load_method.specialization_failure_kinds[18] : 416 (class method)`
If so, then we should be able to get near 100% hits with specialization for class and module functions.
Which would be fantastic.