@Fidget-Spinner 2021-08-12
```
python.exe -m test test_typing test_re test_dis test_zlib

LOAD_METHOD_CACHED, LOAD_METHOD_MODULE, LOAD_METHOD_CLASS
    load_method.specialization_success : 1357
    load_method.specialization_failure : 917
    load_method.hit : 945963
    load_method.deferred : 66527
    load_method.miss : 1420
    load_method.deopt : 191
    load_method.unquickened : 8639
    load_method.specialization_failure_kinds[7] : 1
    load_method.specialization_failure_kinds[17] : 33  (builtin classmethod)
    load_method.specialization_failure_kinds[18] : 15  (@classmethod obj)
    load_method.specialization_failure_kinds[19] : 431 (not a method, maybe your 7%?)
```
``LOAD_METHOD_CLASS`` might not be worth adding because it makes every ``LOAD_METHOD`` optimization site require 8 more bytes for a 10-15% increase in cache hits. (I required 1 more cache to keep track of the type's address). I don't know of any other way to verify that something is a valid ``PyTypeObject``. If you could think of something, that'd be great.

Either way, the original specialization (LOAD_METHOD_CACHED) already paid for the specialization in pyperformance itself.