@Fidget-Spinner 2021-08-12
Hmm, I think this line is causing ASAN to fail https://github.com/Fidget-Spinner/cpython/blob/specialize_load_method/Python/specialize.c#L838.
`cache1->tp_version = ((PyTypeObject *)owner)->tp_version_tag;`

Does anyone know why? `owner` can be either a builtin type (e.g. `dict`, `int`), or a class. Is casting owner to `PyTypeObject` pointer unsafe?