@Fidget-Spinner 2021-08-12
I'm surprised that this was such a good specialization candidate. >92% of `LOAD_METHOD` at runtime are now hits from the stats above. This is pretty significant, because a hit make `LOAD_METHOD` nearly free.

Even without `LOAD_METHOD_MODULE` and `LOAD_METHOD_CLASS`, that figure was around 90%.

Using the simple (but maybe flawed) formula: ``100 - (deferred + unquickened) / (deferred + unquickened + hit) * 100``, ``python.exe -m test test_typing test_re test_dis test_zlib`` gives me:

| Op            | %    |
|---------------|------|
| LOAD_GLOBAL   | 98.6 |
| LOAD_METHOD   | 92.6 |
| STORE_ATTR    | 87.4 |
| LOAD_ATTR     | 67.6 |
| BINARY_SUBSCR | 29.9 |

The BINARY_SUBSCR stats are a bit blown out of proportion due to `typing`, since that entire module is custom `__getitem__`.

@markshannon  If you're OK with everything now, all that's left are benchmarks :).