@Fidget-Spinner 2021-08-13
Before I commit the refactoring changes, here are one more set of pyperformance benchmarks:

```
Slower (3):
- pickle_list: 9.95 us +- 0.15 us -> 10.6 us +- 2.5 us: 1.07x slower
- scimark_sparse_mat_mult: 13.1 ms +- 0.2 ms -> 13.5 ms +- 0.5 ms: 1.03x slower
- scimark_lu: 386 ms +- 13 ms -> 394 ms +- 7 ms: 1.02x slower

Faster (24):
- logging_silent: 395 ns +- 10 ns -> 334 ns +- 6 ns: 1.18x faster
- deltablue: 14.8 ms +- 0.2 ms -> 13.4 ms +- 0.2 ms: 1.10x faster
- logging_format: 19.5 us +- 0.4 us -> 17.9 us +- 0.3 us: 1.09x faster
- logging_simple: 17.7 us +- 0.3 us -> 16.3 us +- 0.2 us: 1.09x faster
- pickle_pure_python: 1.05 ms +- 0.04 ms -> 980 us +- 8 us: 1.07x faster
- raytrace: 993 ms +- 11 ms -> 929 ms +- 9 ms: 1.07x faster
- mako: 35.1 ms +- 0.3 ms -> 32.9 ms +- 0.2 ms: 1.07x faster
- hexiom: 21.3 ms +- 0.2 ms -> 20.1 ms +- 0.2 ms: 1.06x faster
- pyflate: 1.55 sec +- 0.01 sec -> 1.46 sec +- 0.01 sec: 1.06x faster
- richards: 157 ms +- 3 ms -> 150 ms +- 2 ms: 1.04x faster
- scimark_sor: 481 ms +- 6 ms -> 461 ms +- 3 ms: 1.04x faster
- chaos: 233 ms +- 2 ms -> 224 ms +- 2 ms: 1.04x faster
- pickle_dict: 63.2 us +- 0.6 us -> 60.7 us +- 6.8 us: 1.04x faster
- json_dumps: 32.1 ms +- 0.5 ms -> 30.9 ms +- 0.4 ms: 1.04x faster
- tornado_http: 359 ms +- 6 ms -> 346 ms +- 7 ms: 1.04x faster
- scimark_monte_carlo: 235 ms +- 3 ms -> 227 ms +- 2 ms: 1.04x faster
- unpickle_pure_python: 715 us +- 7 us -> 692 us +- 6 us: 1.03x faster
- chameleon: 21.5 ms +- 0.2 ms -> 20.9 ms +- 0.2 ms: 1.03x faster
- django_template: 107 ms +- 1 ms -> 104 ms +- 2 ms: 1.03x faster
- xml_etree_process: 168 ms +- 3 ms -> 164 ms +- 3 ms: 1.03x faster
- scimark_fft: 1.05 sec +- 0.01 sec -> 1.03 sec +- 0.01 sec: 1.03x faster
- sqlite_synth: 7.02 us +- 0.47 us -> 6.84 us +- 0.15 us: 1.03x faster
- dulwich_log: 210 ms +- 3 ms -> 205 ms +- 4 ms: 1.02x faster
- go: 464 ms +- 5 ms -> 454 ms +- 5 ms: 1.02x faster

Benchmark hidden because not significant (31): 2to3, crypto_pyaes, fannkuch, float, json_loads, meteor_contest, nbody, nqueens, pathlib, pickle, pidigits, python_startup, python_startup_no_site, regex_compile, regex_dna, regex_effbot, regex_v8, spectral_norm, sqlalchemy_declarative, sqlalchemy_imperative, sympy_expand, sympy_integrate, sympy_sum, sympy_str, telco, unpack_sequence, unpickle, unpickle_list, xml_etree_parse, xml_etree_iterparse, xml_etree_generate

Geometric mean: 1.02x faster
```